Progressive Rendering
Angler now supports progressive rendering

+++ AnglerRT/includes/core/ImageHandler.h
@@ -8,7 +8,7 @@
 class Image {
   public:
     virtual void Write(const Color &) = 0;
-    virtual std::unique_ptr<uint8_t[]> getBufferCopy() const = 0;
+    virtual std::unique_ptr<float[]> getBufferCopy() const = 0;
     virtual unsigned getHeight() const = 0;
     virtual unsigned getWidth() const = 0;
 };
